+++ open-xchange-appsuite-public-sector/vite.config.js
@@ -128,7 +128,11 @@ export default defineConfig(({ mode }) => {
       // copy existing assets from source directory that are not imported in source code
       rollupPluginCopy({
         targets: [
-          { src: './src/io.ox.public-sector/*.svg', dest: 'public/io.ox.public-sector/' }
+          { src: './src/io.ox.public-sector/*.svg', dest: 'public/io.ox.public-sector/' },
+          {
+            src: './src/io.ox.public-sector/navigation/*.svg',
+            dest: 'public/io.ox.public-sector/navigation/'
+          }
         ],
         hook: 'buildStart'
       })