+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- PBSR-640: Upgrade central navigation to OX 8
 
 ## [1.0.0] - 2023-05-09
 - PBSR-576: Ported customizations to OX 8